restriction.[ch]pp: Harmonize types per warnings.
Explicitly assume that T(Signed)SeqPos will suffice for sequence data,
with the help of a new CPatternRec::GetPatternSize(), but (formally?)
allow full size_t for some other quantities, such as the number of
expansions of a sequence with ambiguity characters.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@103392 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/sequence/restriction.hpp
@@ -519,9 +519,9 @@ private:
                            CRebase::EEnzymesToLoad which_enzymes);
 
     static void x_ExpandRecursion(string& s, unsigned int pos,
-                                  CTextFsm<int>& fsm, int match_value);
-    static void x_AddPattern(const string& pat, CTextFsm<int>& fsm,
-                             int match_value);
+                                  CTextFsm<size_t>& fsm, size_t match_value);
+    static void x_AddPattern(const string& pat, CTextFsm<size_t>& fsm,
+                             size_t match_value);
     static bool x_IsAmbig(char nuc);
 
     template<class Seq>
